Rename graphs to plots
