+++ docs/app/guides/display/plots/page.tsx
@@ -7,22 +7,22 @@ import SineSource from "!raw-loader!guide-examples/hello-fx/sine.tsx"
 import Parametric from "guide-examples/Parametric"
 import ParametricSource from "!raw-loader!guide-examples/Parametric.tsx"
 import { PropTable } from "components/PropTable"
-import { FunctionGraph } from "mafs"
+import { Plot } from "mafs"
 
 function Functions() {
   return (
     <>
-      {/* <h2>
+      <h2>
         Functions of <em className="font-display">x</em>
-      </h2> */}
+      </h2>
 
       <CodeAndExample component={<Sine />} source={SineSource} />
-      <PropTable of={FunctionGraph.OfX} displayName="FunctionGraph.OfX" />
+      <PropTable of={Plot.OfX} displayName="FunctionGraph.OfX" />
 
       <h2>Parametric functions</h2>
 
       <CodeAndExample component={<Parametric />} source={ParametricSource} />
-      <PropTable of={FunctionGraph.Parametric} displayName="FunctionGraph.Parametric" />
+      <PropTable of={Plot.Parametric} displayName="FunctionGraph.Parametric" />
     </>
   )
 }
