+++ docs/components/guide-examples/Parametric.tsx
@@ -3,7 +3,7 @@
 import {
   Mafs,
   CartesianCoordinates,
-  FunctionGraph,
+  Plot,
   Theme,
 } from "mafs"
 
@@ -11,7 +11,7 @@ export default function Parametric() {
   return (
     <Mafs height={300} viewBox={{ x: [-3, 3], y: [-3, 3] }}>
       <CartesianCoordinates />
-      <FunctionGraph.Parametric
+      <Plot.Parametric
         xy={(t) => [
           (t / 4) * Math.cos(t * 2 * Math.PI),
           (t / 4) * Math.sin(t * 2 * Math.PI),
