+++ docs/components/guide-examples/hello-fx/sine.tsx
@@ -1,25 +1,12 @@
 "use client"
 
-import {
-  Mafs,
-  CartesianCoordinates,
-  FunctionGraph,
-  Theme,
-} from "mafs"
+import { Mafs, CartesianCoordinates, Plot } from "mafs"
 
 export default function HelloFx() {
   return (
-    <div>
-      <Mafs height={300} viewBox={{ y: [-5, 5] }}>
-        <CartesianCoordinates subdivisions={2} />
-        <FunctionGraph.Parametric
-          xy={(t) => [t, Math.tan(t)]}
-          t={[-10, 10]}
-          minimumSamplingDepth={12}
-          color={Theme.blue}
-          weight={1}
-        />
-      </Mafs>
-    </div>
+    <Mafs height={300}>
+      <CartesianCoordinates subdivisions={4} />
+      <Plot.OfX y={(x) => Math.sin(x)} />
+    </Mafs>
   )
 }
