--- src/display/FunctionGraph.tsx
@@ -1,11 +1,0 @@
-import { OfX } from "./FunctionGraph/OfX"
-import { ParametricFunction as Parametric } from "./FunctionGraph/Parametric"
-
-export const FunctionGraph = {
-  OfX,
-  Parametric,
-}
-
-import type { OfXProps } from "./FunctionGraph/OfX"
-import type { ParametricProps } from "./FunctionGraph/Parametric"
-export type { OfXProps, ParametricProps }
