+++ src/display/Plot.tsx
@@ -1,0 +1,11 @@
+import { OfX } from "./Plot/FunctionGraph/OfX"
+import { ParametricFunction as Parametric } from "./Plot/Parametric"
+
+export const Plot = {
+  OfX,
+  Parametric,
+}
+
+import type { OfXProps } from "./Plot/FunctionGraph/OfX"
+import type { ParametricProps } from "./Plot/Parametric"
+export type { OfXProps, ParametricProps }
